fix: change mode set cursor
